backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
    #   starlette
argon2-cffi==25.1.0
    # via minio
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
bcrypt==5.0.0
    # via -r requirements.in
billiard==4.2.4
    # via celery
boto3==1.42.73
    # via -r requirements.in
botocore==1.42.73
    # via
    #   boto3
    #   s3transfer
celery==5.6.2
    # via
    #   -r requirements.in
    #   flower
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   minio
    #   requests
cffi==2.0.0
    # via argon2-cffi-bindings
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via pydantic
fastapi==0.135.1
    # via -r requirements.in
flower==2.0.1
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
humanize==4.15.0
    # via flower
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jiter==0.13.0
    # via openai
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
kombu==5.6.2
    # via celery
minio==7.2.20
    # via -r requirements.in
openai==2.29.0
    # via -r requirements.in
packaging==26.0
    # via kombu
prometheus-client==0.24.1
    # via flower
prompt-toolkit==3.0.52
    # via click-repl
psycopg[binary,pool]==3.3.3
    # via -r requirements.in
psycopg-binary==3.3.3
    # via psycopg
psycopg-pool==3.3.0
    # via psycopg
pycparser==3.0
    # via cffi
pycryptodome==3.23.0
    # via minio
pydantic[email]==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via -r requirements.in
pyjwt==2.12.1
    # via -r requirements.in
pypdf==6.9.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   pydantic-settings
python-multipart==0.0.22
    # via -r requirements.in
pytz==2026.1.post1
    # via flower
redis==7.3.0
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via openai
starlette==0.52.1
    # via fastapi
tornado==6.5.5
    # via flower
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   minio
    #   openai
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via kombu
tzlocal==5.3.1
    # via celery
urllib3==2.6.3
    # via
    #   botocore
    #   minio
    #   requests
uvicorn==0.42.0
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.6.0
    # via prompt-toolkit
